+++ lib/g5_integrations_validations/yardi_rent_cafe_inventory_rate_to_display_validations.rb
@@ -1,0 +1,17 @@
+module G5IntegrationsValidations
+  module YardiRentCafeInventoryRateToDisplayValidations
+    extend ActiveSupport::Concern
+
+    VALID_OPTIONS = %w(starting_rate starting_and_ending_rates none)
+
+    included do
+      include ActiveModel::Validations
+
+      validates(:rate_to_display, {
+        inclusion: {in: VALID_OPTIONS},
+        presence: true,
+      })
+    end
+  end
+
+end