+++ spec/fixtures/yardi_rent_cafe_inventory.rb
@@ -1,0 +1,8 @@
+class YardiRentCafeInventory
+
+  include G5IntegrationsValidations::YardiVoyagerInventoryRateToDisplayValidations
+  include Virtus.model
+
+  attribute :rate_to_display, String
+
+end